refactor : `DBColumnMapper` 인터페이스화를 통한 결합도 낮추기
